requirements.txt
alembic==1.14.0
biopython==1.84
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.0
packaging==24.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3